+++ .gitignore
@@ -1,0 +1,4 @@
+/.idea/
+/build/
+/dist/
+/*.egg-info/
