+++ scanpy_cli/scanpy_filter_cells.py
@@ -1,12 +1,23 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 
-from __future__ import print_function
+import sys
 import logging
 import scanpy.api as sc
-from scanpy_wrapper_utils import ScanpyArgParser, read_input_object, write_output_object
+
+from scanpy_cli.wrapper_utils import (
+    ScanpyArgParser,
+    read_input_object, write_output_object,
+)
 
 
-def main(args):
+def main(argv=None):
+    argparser = ScanpyArgParser(argv, 'Filter cells by properties and/or simple stats')
+    argparser.add_input_object()
+    argparser.add_output_object()
+    argparser.add_subset_parameters()
+    argparser.add_subset_list()
+    args = argparser.args
+
     logging.debug(args)
 
     adata = read_input_object(args.input_object_file, args.input_format)
@@ -39,11 +50,4 @@ def main(args):
 
 
 if __name__ == "__main__":
-    argparser = ScanpyArgParser('Filter cells by properties and/or simple stats')
-    argparser.add_input_object()
-    argparser.add_output_object()
-    argparser.add_subset_parameters()
-    argparser.add_subset_list()
-    args = argparser.get_args()
-
-    main(args)
+    main()
