+++ scanpy_cli/scanpy_find_cluster.py
@@ -1,39 +1,15 @@
 #!/usr/bin/env python
 
-from __future__ import print_function
 import logging
-from scanpy_wrapper_utils import ScanpyArgParser, comma_separated_list
-from scanpy_wrapper_utils import read_input_object, write_output_object
+
+from scanpy_cli.wrapper_utils import (
+    ScanpyArgParser, comma_separated_list,
+    read_input_object, write_output_object,
+)
 
 
-def main(args):
-    logging.debug(args)
-    import scanpy.api as sc
-
-    adata = read_input_object(args.input_object_file, args.input_format)
-
-    if args.restrict_to is not None:
-        args.restrict_to = (args.restrict_to[0], args.restrict_to[1:])
-    sc.tl.louvain(adata,
-                  flavor=args.flavor,
-                  resolution=args.resolution,
-                  restrict_to=args.restrict_to,
-                  key_added=args.key_added,
-                  use_weights=args.use_weights,
-                  random_state=args.random_seed)
-
-    write_output_object(adata, args.output_object_file, args.output_format)
-
-    if args.output_text_file:
-        adata.obs[[args.key_added]].reset_index(level=0).rename(columns={'index':'cells'}).to_csv(
-            args.output_text_file, sep='\t', index=None)
-
-    logging.info('Done')
-    return 0
-
-
-if __name__ == '__main__':
-    argparser = ScanpyArgParser('Run Louvain clustering on data with neighborhood graph computed')
+def main(argv=None):
+    argparser = ScanpyArgParser(argv, 'Run Louvain clustering on data with neighborhood graph computed')
     argparser.add_input_object()
     argparser.add_output_object()
     argparser.add_argument('--output-text-file',
@@ -67,6 +43,32 @@ if __name__ == '__main__':
                            type=int,
                            default=0,
                            help='The seed used to initialise optimisation. Default: 0')
-    args = argparser.get_args()
+    args = argparser.args
 
-    main(args)
+    logging.debug(args)
+    import scanpy.api as sc
+
+    adata = read_input_object(args.input_object_file, args.input_format)
+
+    if args.restrict_to is not None:
+        args.restrict_to = (args.restrict_to[0], args.restrict_to[1:])
+    sc.tl.louvain(adata,
+                  flavor=args.flavor,
+                  resolution=args.resolution,
+                  restrict_to=args.restrict_to,
+                  key_added=args.key_added,
+                  use_weights=args.use_weights,
+                  random_state=args.random_seed)
+
+    write_output_object(adata, args.output_object_file, args.output_format)
+
+    if args.output_text_file:
+        adata.obs[[args.key_added]].reset_index(level=0).rename(columns={'index':'cells'}).to_csv(
+            args.output_text_file, sep='\t', index=None)
+
+    logging.info('Done')
+    return 0
+
+
+if __name__ == '__main__':
+    main()
