+++ scanpy_cli/scanpy_find_variable_genes.py
@@ -1,17 +1,39 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 
-from __future__ import print_function
 import logging
 import matplotlib
 matplotlib.use('Agg')
+
 import scanpy.api as sc
-from scanpy_wrapper_utils import ScanpyArgParser
-from scanpy_wrapper_utils import read_input_object
-from scanpy_wrapper_utils import write_output_object
-from scanpy_wrapper_utils import save_output_plot
+from scanpy_cli.wrapper_utils import (
+    ScanpyArgParser,
+    read_input_object, write_output_object, save_output_plot,
+)
 
 
-def main(args):
+def main(argv=None):
+    argparser = ScanpyArgParser(argv, 'Keep variable genes')
+    argparser.add_input_object()
+    argparser.add_output_object()
+    argparser.add_output_plot()
+    argparser.add_argument('--flavor',
+                           choices=['seurat', 'cell_ranger'],
+                           default='seurat',
+                           help='Choose the flavor for computing normalised dispersion. '
+                                'Default: seurat')
+    argparser.add_subset_parameters(params=['mean', 'disp'])
+    argparser.add_argument('-b', '--n-bins',
+                           type=int,
+                           default=20,
+                           help='Number of bins for binning the mean gene expression. '
+                                'Normalisation is done with respect to each bin. Default: 20')
+    argparser.add_argument('-n', '--n-top-genes',
+                           type=int,
+                           default=None,
+                           help='Number of highly variable genes to keep, '
+                                'ignoring --subset-parameters when set. Default: None')
+    args = argparser.args
+
     logging.debug(args)
 
     adata = read_input_object(args.input_object_file, args.input_format)
@@ -56,26 +78,4 @@ def main(args):
 
 
 if __name__ == '__main__':
-    argparser = ScanpyArgParser('Keep variable genes')
-    argparser.add_input_object()
-    argparser.add_output_object()
-    argparser.add_output_plot()
-    argparser.add_argument('--flavor',
-                           choices=['seurat', 'cell_ranger'],
-                           default='seurat',
-                           help='Choose the flavor for computing normalised dispersion. '
-                                'Default: seurat')
-    argparser.add_subset_parameters(params=['mean', 'disp'])
-    argparser.add_argument('-b', '--n-bins',
-                           type=int,
-                           default=20,
-                           help='Number of bins for binning the mean gene expression. '
-                                'Normalisation is done with respect to each bin. Default: 20')
-    argparser.add_argument('-n', '--n-top-genes',
-                           type=int,
-                           default=None,
-                           help='Number of highly variable genes to keep, '
-                                'ignoring --subset-parameters when set. Default: None')
-    args = argparser.get_args()
-
-    main(args)
+    main()
