+++ scanpy_cli/scanpy_run_umap.py
@@ -1,60 +1,18 @@
 #!/usr/bin/env python
 
-from __future__ import print_function
 import logging
 import matplotlib
 matplotlib.use('Agg')
-from scanpy_wrapper_utils import ScanpyArgParser, comma_separated_list
-from scanpy_wrapper_utils import read_input_object, write_output_object, save_output_plot
+import scanpy.api as sc
+
+from scanpy_cli.wrapper_utils import (
+    ScanpyArgParser,
+    read_input_object, write_output_object, save_output_plot,
+)
 
 
-def main(args):
-    logging.debug(args)
-    import scanpy.api as sc
-
-    adata = read_input_object(args.input_object_file, args.input_format)
-
-    sc.tl.umap(adata,
-               min_dist=args.min_dist,
-               spread=args.spread,
-               n_components=args.n_components,
-               maxiter=args.maxiter,
-               alpha=args.alpha,
-               gamma=args.gamma,
-               negative_sample_rate=args.negative_sample_rate,
-               init_pos=args.init_pos,
-               random_state=args.random_seed,
-               a=args.a,
-               b=args.b)
-
-    write_output_object(adata, args.output_object_file, args.output_format)
-
-    if args.output_embeddings_file:
-        adata.obsm.to_df()[['X_umap1', 'X_umap2']].to_csv(args.output_embeddings_file, index=None)
-
-    if args.output_plot:
-        sc.set_figure_params(format=args.output_plot_format)
-        sc.settings.verbosity = 0
-        sc.pl.umap(adata, show=False, save=True,
-                   color=args.color_by,
-                   use_raw=args.use_raw,
-                   edges=args.edges,
-                   arrows=args.arrows,
-                   sort_order=args.sort_order,
-                   groups=args.groups,
-                   projection=args.projection,
-                   components=args.components,
-                   palette=args.palette,
-                   frameon=args.frameon)
-
-        save_output_plot('umap', args.output_plot)
-
-    logging.info('Done')
-    return 0
-
-
-if __name__ == '__main__':
-    argparser = ScanpyArgParser('Run UMAP on data with neighborhood graph computed')
+def main(argv=None):
+    argparser = ScanpyArgParser(argv, 'Run UMAP on data with neighborhood graph computed')
     argparser.add_input_object()
     argparser.add_output_object()
     argparser.add_argument('--output-embeddings-file',
@@ -117,6 +75,50 @@ if __name__ == '__main__':
                                 'and --spread. Default: None')
     argparser.add_output_plot()
     argparser.add_scatter_plot_options()
-    args = argparser.get_args()
+    args = argparser.args
 
-    main(args)
+    logging.debug(args)
+
+    adata = read_input_object(args.input_object_file, args.input_format)
+
+    sc.tl.umap(adata,
+               min_dist=args.min_dist,
+               spread=args.spread,
+               n_components=args.n_components,
+               maxiter=args.maxiter,
+               alpha=args.alpha,
+               gamma=args.gamma,
+               negative_sample_rate=args.negative_sample_rate,
+               init_pos=args.init_pos,
+               random_state=args.random_seed,
+               a=args.a,
+               b=args.b)
+
+    write_output_object(adata, args.output_object_file, args.output_format)
+
+    if args.output_embeddings_file:
+        adata.obsm.to_df()[['X_umap1', 'X_umap2']].to_csv(args.output_embeddings_file, index=None)
+
+    if args.output_plot:
+        sc.set_figure_params(format=args.output_plot_format)
+        sc.settings.verbosity = 0
+        sc.pl.umap(adata, show=False, save=True,
+                   color=args.color_by,
+                   use_raw=args.use_raw,
+                   edges=args.edges,
+                   arrows=args.arrows,
+                   sort_order=args.sort_order,
+                   groups=args.groups,
+                   projection=args.projection,
+                   components=args.components,
+                   palette=args.palette,
+                   frameon=args.frameon)
+
+        save_output_plot('umap', args.output_plot)
+
+    logging.info('Done')
+    return 0
+
+
+if __name__ == '__main__':
+    main()
